Add Manga resource and creation
This adds creation of Manga records. This also adds the 2 big things:

- Concept directory
- Content directory

To explain, code will be categorized into concepts where each concept
directory will have its own set of directories pertaining to the various
requirements of the Commanded library.

For this commit, the Content concept is added which ideally should
contain code related to content creation on the app; Namely, Manga.